+++ nddata/utils/__init__.py
@@ -2,3 +2,4 @@ from . import decorators
 from . import garbagecollector
 from . import descriptors
 from . import numpyutils
+from . import sentinels